Note possibility of a release branch cut before a release
Grafana v9.1 cut the `v9.1.x` branch two weeks before the final v9.1 release. This created an edge case in the docs publishing and Grafana release process that resulted in docs content and changes being merged against the `main` branch but not being published in the `/latest/` docs when v9.1 was released. Those docs PRs needed to be backported to the `v9.1.x` branch even if Grafana v9.1 had not been released and the v9.1 docs had not yet been published to `/latest/`.

Note this and adjust the workflow diagram accordingly.