Add make-docs script that can be used to build multiple documentation sets
With the script served from the writers-toolkit documentation, it can
be sourced by any repository and used everywhere.

The simplifications to the Makefile eliminate the need to update every
repository when a new feature is added to the local webserver
workflow.

The script encapsulates all the best practices we have established
across the various `make docs` workflows, including:

1. Using `--init` flag to ensure that ^C works on MacOS.
1. Using `--platform linux/amd64` to ensure virtualization on MacOS.
1. Using `podman` instead of `docker` if the command is available.
1. Implementing a redirect from the project root to the mounted
version.
1. Logging the webserver URLs for convenient browsing.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>