Use PHONY update target to avoid causing diffs in CI
CI typically invokes `make -B` because git doesn't store timestamps.
This results in changes to the working tree when invoking `make docs` if the version of `docs.mk` committed to the repo is different to one in Writers' Toolkit.
Some CI tests use the exit status of `git diff` to determine if generated files have been committed and this working tree change interferes.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>