Fix instructions for placeholder variables
The example demonstrated Markdown used in a code block which wouldn't be rendered as the example intended.

The preferred Markdown syntax for italic emphasis is `_` when using `prettier` so I've updated the instructions to use that instead of `*`.

I've also fixed the name of the characters to match their unicode naming and added examples of each of them in case the name is not obvious to readers

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>