Review "Links" documentation
In its previous form, readers were confused into believing the only behavior of version substitution was front matter lookup which provides little value over linking without the syntax.

The rewrite intends to more clearly state that the default behavior is inference from the current page but that it can be overridden with front matter.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>